fix image locations
images/ has to be in docs/ to show up in live site.
We also want images to show up when previewing.
But the image location in the html must be the same in preview/ and docs/.
I think easiest way to achieve this is just to copy images/ in both
preview/ and docs/.